LayoutTests/css3/flexbox/bug646288.html

<!DOCTYPE html>
<style>
body { width: 800px; }

* { min-width: 0; min-height: 0; }

#list-wrapper {
  display: flex;
  flex-direction: column;
}

#player-list-wrapper {
  flex-grow: 1;
  overflow: auto;
}
</style>

<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>

<p>Tests that there's no vertical scrollbar for this content</p>

<body onload="checkLayout('#list-wrapper')">
  <div id="list-wrapper" data-expected-width="800">
    <div id="player-list-wrapper" data-expected-width="800">
      <div data-expected-width="800">http://www.ispeech.org/p/generic/getaudio?text=http%3A%2F%2Fwww.ispeech.org%2Ftext.to.speech%3Flink%3Dhttp%253A%252F%252Fwww.ispeech.org%252Ftext.to.speech%253Fvoice%253Dusenglishfemale%2526action%253Dconvert%2526speed%253D0%2526text%253DType%252520some%252520text%252520and%252520click%252520play%2C&amp;voice=usenglishfemale&amp;speed=0&amp;action=convert</div>
    </div>
  </div>

</body>
